requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
flask==3.1.0
numpy==1.26.4
torch==2.2.0
Pillow==10.2.0
